Update the wrapper for the plotting macro

+++ Fitter/scripts/overlay_operator_variations_wrapper.py
@@ -2,11 +2,11 @@ import os
 import logging
 import subprocess
 import datetime
+import argparse
 
 import EFTFit.Fitter.consts as CONST
 
 from EFTFit.Fitter.LvlFormatter import LvlFormatter
-#from EFTFit.Fitter.CombineHelper import CombineHelper
 from EFTFit.Fitter.utils import get_files, move_files, clean_dir
 from EFTFit.Fitter.make_html import make_html
 
@@ -15,6 +15,22 @@ from EFTFit.Fitter.make_html import make_html
 #   by the combine_helper.py code. It looks for input directories starting in the 'test' directory
 #   of the EFTFit/Fitter package and tries to create and move output files (usually png/pdf plots)
 #   to a webarea starting in ~/www/eft_stuff/misc/fitting_plots
+
+parser = argparse.ArgumentParser(description='Customize the plotting macro')
+parser.add_argument('--inpath',  '-i', default=os.getcwd()   , help = 'Input directory')
+parser.add_argument('--outpath', '-o', default=None          , help = 'Output directory')
+parser.add_argument('--test',   '-t' , action='store_true'   , help = 'Run test')
+
+args = parser.parse_args()
+in_dir  = args.inpath
+out_dir = args.outpath
+dotest  = args.test
+
+if dotest:
+    if out_dir is None:
+        out_dir = os.path.join(os.getcwd(), 'test')
+    else:
+        out_dir = os.path.join(out_dir, 'test')
 
 TSTAMP1 = datetime.datetime.now().strftime('%Y-%m-%d_%H%M')
 TSTAMP2 = datetime.datetime.now().strftime('%Y-%m-%d')
@@ -28,50 +44,15 @@ console.setLevel(logging.INFO)
 console.setFormatter(frmt)
 logging.getLogger('').addHandler(console)
 
-#helper = CombineHelper(out_dir="test_overlay_wrapper")
-
-web_area = os.path.join(CONST.USER_DIR,'www')
-dir_ver  = 'data2017v1'  # This is sort of versioning for which version of the histogram file is being used
-testing  = True 
-
-dirs = [
-    #'SM_fitting_2019-04-04_1051_simultaneous_allcats',
-    #'SM_fitting_2019-04-04_1945_simultaneous_3l',
-    #'SM_fitting_2019-04-03_1357_simultaneous-notllq_2lss'
-    #'SM_fitting_2019-04-10_2022_simultaneous_adhocSFs_allcats',
-    #'SM_fitting_2019-04-10_2144_simultaneous_allcats_centralSamples',
-    #'SM_fitting_2019-04-19_1229_simultaneous_reducedJetBins_allcats',
-    # 'SM_fitting_2019-04-23_1822_simultaneous_newAdhocSFs_allcats'
-    # 'SM_fitting_2019-12-06_ana32_private'
-    # 'SM_fitting_2020-01-10_ana32_central'
-    # 'SM_fitting_2020-01-11_ana32_central'
-    # 'SM_fitting_2020-01-22_ana32_private_sgnl_Default-PSWeights-Symmetrized_MatchingSyst_moreStats'
-    # 'SM_fitting_2020-10-13_ana32_private_sgnl_Decorrelated-PS-PDF-Q2RF-QCUT_with-NuisOnly_cmin0',   # SM signal strength fit (for supplementary plots)
-    # 'EFT_fitting_testing'
-    # 'EFT_fitting_2020-01-10_ana32_private'
-    # 'EFT_fitting_2020-01-22_ana32_private_sgnl_Default-PSWeights-Symmetrized_MatchingSyst_moreStats'
-    # 'EFT_fitting_2020-03-31_ana32_private_sgnl_Default-PSWeights-Symmetrized_MatchingSyst_moreStats_with-NuisOnly',
-    # 'EFT_fitting_2020-04-01_ana32_private_sgnl_Decorrelated-PS-PDF-Q2RF-QCUT_with-NuisOnly',
-    'EFT_fitting_2020-04-02_ana32_private_sgnl_Decorrelated-PS-PDF-Q2RF-QCUT_with-NuisOnly_cmin0', # This is the one most commonly used in the TOP-19-001 Paper
-    # 'EFT_fitting_2020-04-02_ana32_private_sgnl_Decorrelated-PS-PDF-Q2RF-QCUT_with-NuisOnly_cmin2',
-    # 'EFT_fitting_2020-09-04_ana32_private_sgnl_Decorrelated-PS-PDF-Q2RF-QCUT_with-NuisOnly_no-SFZ-bin-mask_cmin0',
-
-]
-
-for d in dirs:
-    #print d
-    input_dir  = os.getcwd()#os.path.join(CONST.EFTFIT_TEST_DIR,dir_ver,d)
-    output_dir = os.path.join(web_area,'EFT/fitting_plots',dir_ver,"{tstamp}_{dir}".format(tstamp=TSTAMP2,dir=d))
-    if testing:
-        output_dir = os.path.join(web_area,'EFT/fitting_plots','testing')
-    # output_dir = helper.getOutputDirectory()
-
-    if not os.path.exists(output_dir):
-        print "Making output directory: %s" % (output_dir)
-        os.makedirs(output_dir)
-    #helper.cleanDirectory(output_dir)
-    for filename in os.listdir(output_dir):
-        file_path = os.path.join(output_dir, filename)
+# Delete the files on the output directory only if it's given by the user
+if out_dir is None:
+    out_dir = os.getcwd() # Default output path: the current directory
+else:
+    if not os.path.exists(out_dir):
+        print "Making output directory: %s" % (out_dir)
+        os.makedirs(out_dir)
+    for filename in os.listdir(out_dir):
+        file_path = os.path.join(out_dir, filename)
         try:
             if os.path.isfile(file_path) or os.path.islink(file_path):
                 os.unlink(file_path)
@@ -80,26 +61,27 @@ for d in dirs:
         except Exception as e:
             print('Failed to delete %s. Reason: %s' % (file_path, e))
             
-    macro_dir = os.path.join(CONST.EFTFIT_TEST_DIR,'../scripts/')
-    root_args = "\"{indir}\",\"{outdir}\"".format(indir=input_dir,outdir=output_dir)
-    subprocess.check_call(['root','-b','-l','-q', macro_dir + 'overlay_operator_variations.C({args})'.format(args=root_args)])
+macro_dir = os.path.join(CONST.EFTFIT_TEST_DIR,'../scripts/')
+root_args = "\"{indir}\",\"{outdir}\"".format(indir=in_dir,outdir=out_dir)
+subprocess.check_call(['root','-b','-l','-q', macro_dir + 'overlay_operator_variations.C({args})'.format(args=root_args)])
     
-    to_convert = get_files('.',targets=['.eps','.ps'])
-    for fn in to_convert:
-        continue
-        # Note:
-        #   This is so that the "\ell" symbol can get rendered properly when converted to a PDF and
-        #   displayed in the analysis paper built by CMS TeX tools
-        subprocess.check_call(['sed','-i','-e',"s|STIXGeneral-Italic|STIXXGeneral-Italic|g",fn])
-        # The '-dEPSCrop' is to fix the large whitespace caused by bound box
-        # ps_conv_args = ['ps2pdf14','-dPDFSETTINGS=/prepress','-dEPSCrop',fn]
-        # print " ".join(ps_conv_args)
-        subprocess.check_call(['ps2pdf14','-dPDFSETTINGS=/prepress','-dEPSCrop',fn])
+to_convert = get_files('.',targets=['.eps','.ps'])
+for fn in to_convert:
+    continue
+    # Note:
+    #   This is so that the "\ell" symbol can get rendered properly when converted to a PDF and
+    #   displayed in the analysis paper built by CMS TeX tools
+    subprocess.check_call(['sed','-i','-e',"s|STIXGeneral-Italic|STIXXGeneral-Italic|g",fn])
+    # The '-dEPSCrop' is to fix the large whitespace caused by bound box
+    # ps_conv_args = ['ps2pdf14','-dPDFSETTINGS=/prepress','-dEPSCrop',fn]
+    # print " ".join(ps_conv_args)
+    subprocess.check_call(['ps2pdf14','-dPDFSETTINGS=/prepress','-dEPSCrop',fn])
 
-
+# Skip this step if the output path is the current directory
+if not out_dir is os.getcwd():
     plots = get_files('.',targets=['.png','.pdf','.ps','.eps'])
     if 0:#len(plots):
         for fn in plots:
             print "fname: {}".format(fn)
-        move_files(plots,output_dir)
-        make_html(output_dir)
+        move_files(plots,out_dir)
+        make_html(out_dir)
